Move into lsystem subdir
